Refactor UserWith to simply implementation massively
UserWith no longer needs to know the underlying data send along with the response, as it no longer checks within the extractor itself whenever the user is authorized to perform the operation. Instead the UserWith creates a Policy object which the caller can use to determine authorization. The Policy upon authorization creates a contract which can safely used to perform the operation in a correct manner.